Added magres_workflow to utils due to circular import problem in utils.py
