+++ electronicparsers/utils/__init__.py
@@ -16,6 +16,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .utils import (
-    get_files, numpy_type_to_json_serializable, BeyondDFTWorkflowsParser
-)
+from .utils import get_files, numpy_type_to_json_serializable, BeyondDFTWorkflowsParser
+from .magres_workflow import NMRMagRes, NMRMagResMethod, NMRMagResResults
